feat: move more functions to separate node modules
