+++ lib/constants.js
@@ -1,0 +1,14 @@
+/**
+ * Constants.
+ *
+ * @type {String}
+ */
+const GROUP_DELIMITER = '.';
+const GROUP_MASK = '*';
+const BACKGROUND = 'background';
+const BACKGROUND_IMAGE = 'background-image';
+
+exports.GROUP_DELIMITER = GROUP_DELIMITER;
+exports.GROUP_MASK = GROUP_MASK;
+exports.BACKGROUND = BACKGROUND;
+exports.BACKGROUND_IMAGE = BACKGROUND_IMAGE;
